LayoutTests/fast/canvas/canvas-outside-viewport-timer-throttling.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
var i = 0;
var context;
var timeoutId;
var wasThrottled;

function loop()
{
   i = (i + 1) % 255;
   context.fillStyle = "rgb(" + i + ", 0, 0)";
   context.fillRect(0, 0, 150, 150);

   // 5 iterations should suffice to throttle the timer.
   if (i == 5) {
     debug("The timer should now be throttled as it is drawing on a canvas outside viewport.");
     shouldBeTrue("internals.isTimerThrottled(timeoutId)");
     clearInterval(timeoutId);
     finishJSTest();
   }
}

function runTest() {
  context = document.getElementById("canvas").getContext('2d');

  timeoutId = setInterval(loop, 5);
  debug("The timer should initially not be throttled.");
  shouldBeFalse("internals.isTimerThrottled(timeoutId)");
}
</script>
</head>
<body onload="runTest();">
<script>
description("Test that a DOM timer drawing on a canvas outside viewport gets throttled.");
jsTestIsAsync = true;
</script>
<div style="position: relative; width: 1600px; height: 2400px;">
  <canvas id="canvas" width=150 height=150 style="position:absolute; left: 600px; top: 800px;"></canvas>
</div>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
