LayoutTests/fast/canvas/canvas-outside-viewport-timer-throttling-expected.txt

Test that a DOM timer drawing on a canvas outside viewport gets throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The timer should initially not be throttled.
PASS internals.isTimerThrottled(timeoutId) is false
The timer should now be throttled as it is drawing on a canvas outside viewport.
PASS internals.isTimerThrottled(timeoutId) is true
PASS successfullyParsed is true

TEST COMPLETE
